move treeexec to util
