+++ sympy/rules/tests/test_strat_pure.py
@@ -79,14 +79,6 @@ def test_minimize():
     rl = minimize(inc, dec, objective=lambda x: -x)
     assert rl(4) == 5
 
-def test_treeexec():
-    tree = ([3, 3], [4, 1], 2)
-    assert treeexec(tree, {list: min, tuple: max}) == 3
-
-    add = lambda *args: sum(args)
-    mul = lambda *args: reduce(lambda a, b: a*b, args, 1)
-    assert treeexec(tree, {list: add, tuple: mul}) == 60
-
 def test_treeexec_strategies():
     join = {list: chain, tuple: minimize}
     inc = lambda x: x + 1
